CLIロジックとコアロジックの分離・ユニットテスト追加
Why:
テスト容易性・保守性向上のため、CLIとProvider管理/実行ロジックを分離し、ユニットテストを追加。

What/How:
- myder_core.py新設（Provider管理・実行ロジック集約）
- myder.pyはCLI引数処理のみ担当にリファクタ
- pytest形式のユニットテスト（tests/test_myder_core.py）追加
